fixed THREE major flaws
firstly: not all possessions end in something that has an xTT2 value, so threw in an if-else to account for that. secondly: in the prior code, if a player wasn't a passing player, they didn't get any personal xTT. that has been fixed. finally, if a player wasn't a passing player, they still accumulated some team contribution, though it wasn't added to the final xTT chain, it was just kinda there. it is no longer there.